+++ .formatter.exs
@@ -1,0 +1,13 @@
+locals_without_parens = [
+  project: 2,
+  project: 3
+]
+
+[
+  inputs: [
+    "{mix,.formatter}.exs",
+    "{config,lib,test}/**/*.{ex,exs}"
+  ],
+  locals_without_parens: locals_without_parens,
+  export: [locals_without_parens: locals_without_parens]
+]