+++ config/dev.exs
@@ -1,4 +1,3 @@
 use Mix.Config
 
-config :mix_test_watch,
-  clear: true
+config :mix_test_watch, clear: true