+++ test/projections/deprecated_projection_test.exs
@@ -1,0 +1,67 @@
+defmodule Commanded.Projections.DeprecatedProjectionTest do
+  use ExUnit.Case
+
+  import Commanded.Projections.ProjectionAssertions
+  import ExUnit.CaptureIO
+
+  alias Commanded.Projections.Repo
+
+  defmodule AnEvent do
+    defstruct name: "AnEvent"
+  end
+
+  defmodule Projection do
+    use Ecto.Schema
+
+    schema "projections" do
+      field(:name, :string)
+    end
+  end
+
+  setup do
+    Ecto.Adapters.SQL.Sandbox.checkout(Repo)
+  end
+
+  test "should handle a projected event" do
+    capture_io(:stderr, fn ->
+      defmodule DeprecatedProjector do
+        use Commanded.Projections.Ecto, name: "DeprecatedProjector"
+
+        project %AnEvent{name: name}, _metadata do
+          Ecto.Multi.insert(multi, :my_projection, %Projection{name: name})
+        end
+      end
+
+      assert :ok == DeprecatedProjector.handle(%AnEvent{}, %{event_number: 1})
+
+      assert_projections(Projection, ["AnEvent"])
+      assert_seen_event("DeprecatedProjector", 1)
+    end)
+  end
+
+  test "should warn project/2 macro deprecated" do
+    assert capture_io(:stderr, fn ->
+             defmodule DeprecatedProjectorWarn2 do
+               use Commanded.Projections.Ecto, name: "DeprecatedProjectorWarn2"
+
+               project %AnEvent{name: name} do
+                 Ecto.Multi.insert(multi, :my_projection, %Projection{name: name})
+               end
+             end
+           end) =~
+             "project macro with \"do end\" block is deprecated; use project/2 with function instead"
+  end
+
+  test "should warn project/3 macro deprecated" do
+    assert capture_io(:stderr, fn ->
+             defmodule DeprecatedProjectorWarn3 do
+               use Commanded.Projections.Ecto, name: "DeprecatedProjectorWarn3"
+
+               project %AnEvent{name: name}, _metadata do
+                 Ecto.Multi.insert(multi, :my_projection, %Projection{name: name})
+               end
+             end
+           end) =~
+             "project macro with \"do end\" block is deprecated; use project/3 with function instead"
+  end
+end