+++ test/projections/ecto_projection_test.exs
@@ -1,20 +1,31 @@
-defmodule Commanded.Projections.EctoTest do
+defmodule Commanded.Projections.EctoProjectionTest do
   use ExUnit.Case
 
   import Commanded.Projections.ProjectionAssertions
 
   alias Commanded.Projections.Repo
 
-  defmodule AnEvent, do: defstruct [name: "AnEvent"]
-  defmodule AnotherEvent, do: defstruct [name: "AnotherEvent"]
-  defmodule IgnoredEvent, do: defstruct [name: "IgnoredEvent"]
-  defmodule ErrorEvent, do: defstruct [name: "ErrorEvent"]
+  defmodule AnEvent do
+    defstruct name: "AnEvent"
+  end
+
+  defmodule AnotherEvent do
+    defstruct name: "AnotherEvent"
+  end
+
+  defmodule IgnoredEvent do
+    defstruct name: "IgnoredEvent"
+  end
+
+  defmodule ErrorEvent do
+    defstruct name: "ErrorEvent"
+  end
 
   defmodule Projection do
     use Ecto.Schema
 
     schema "projections" do
-      field :name, :string
+      field(:name, :string)
     end
   end
 
@@ -41,7 +52,7 @@ defmodule Commanded.Projections.EctoTest do
   test "should handle a projected event" do
     assert :ok == Projector.handle(%AnEvent{}, %{event_number: 1})
 
-    assert_projections Projection, ["AnEvent"]
+    assert_projections(Projection, ["AnEvent"])
     assert_seen_event("Projector", 1)
   end
 
@@ -49,7 +60,7 @@ defmodule Commanded.Projections.EctoTest do
     assert :ok == Projector.handle(%AnEvent{}, %{event_number: 1})
     assert :ok == Projector.handle(%AnotherEvent{}, %{event_number: 2})
 
-    assert_projections Projection, ["AnEvent", "AnotherEvent"]
+    assert_projections(Projection, ["AnEvent", "AnotherEvent"])
     assert_seen_event("Projector", 2)
   end
 
@@ -58,14 +69,14 @@ defmodule Commanded.Projections.EctoTest do
     assert :ok == Projector.handle(%AnEvent{}, %{event_number: 1})
     assert :ok == Projector.handle(%AnEvent{}, %{event_number: 1})
 
-    assert_projections Projection, ["AnEvent"]
+    assert_projections(Projection, ["AnEvent"])
     assert_seen_event("Projector", 1)
   end
 
   test "should ignore unprojected event" do
     assert :ok == Projector.handle(%IgnoredEvent{}, %{event_number: 1})
 
-    assert_projections Projection, []
+    assert_projections(Projection, [])
   end
 
   test "should ignore unprojected events amongst projections" do
@@ -74,14 +85,14 @@ defmodule Commanded.Projections.EctoTest do
     assert :ok == Projector.handle(%AnotherEvent{}, %{event_number: 3})
     assert :ok == Projector.handle(%IgnoredEvent{}, %{event_number: 4})
 
-    assert_projections Projection, ["AnEvent", "AnotherEvent"]
+    assert_projections(Projection, ["AnEvent", "AnotherEvent"])
     assert_seen_event("Projector", 3)
   end
 
   test "should return an error on failure" do
     assert {:error, :failure} == Projector.handle(%ErrorEvent{}, %{event_number: 1})
 
-    assert_projections Projection, []
+    assert_projections(Projection, [])
   end
 
   test "should ensure repo is configured" do
@@ -90,13 +101,15 @@ defmodule Commanded.Projections.EctoTest do
     try do
       Application.put_env(:commanded_ecto_projections, :repo, nil)
 
-      assert_raise RuntimeError, "Commanded Ecto projections expects :repo to be configured in environment", fn ->
-        Code.eval_string """
-        defmodule UnconfiguredProjector do
-          use Commanded.Projections.Ecto, name: "projector"
-        end
-        """
-      end
+      assert_raise RuntimeError,
+                   "Commanded Ecto projections expects :repo to be configured in environment",
+                   fn ->
+                     Code.eval_string("""
+                     defmodule UnconfiguredProjector do
+                       use Commanded.Projections.Ecto, name: "projector"
+                     end
+                     """)
+                   end
     after
       Application.put_env(:commanded_ecto_projections, :repo, repo)
     end
@@ -108,13 +121,13 @@ defmodule Commanded.Projections.EctoTest do
     try do
       Application.put_env(:commanded_ecto_projections, :repo, nil)
 
-      assert Code.eval_string """
-      defmodule ProjectorConfiguredViaOpts do
-        use Commanded.Projections.Ecto,
-          name: "projector",
-          repo: Commanded.Projections.Repo
-      end
-      """
+      assert Code.eval_string("""
+             defmodule ProjectorConfiguredViaOpts do
+               use Commanded.Projections.Ecto,
+                 name: "projector",
+                 repo: Commanded.Projections.Repo
+             end
+             """)
     after
       Application.put_env(:commanded_ecto_projections, :repo, repo)
     end
@@ -122,11 +135,11 @@ defmodule Commanded.Projections.EctoTest do
 
   test "should ensure projection name is present" do
     assert_raise RuntimeError, "UnnamedProjector expects :name to be given", fn ->
-      Code.eval_string """
+      Code.eval_string("""
       defmodule UnnamedProjector do
         use Commanded.Projections.Ecto
       end
-      """
+      """)
     end
   end
 end